lamb

+++ run_pretrain.py
@@ -238,7 +238,7 @@ def model_fn_builder(electra_config, init_checkpoint, learning_rate,
         masked_lm_positions = tf.constant([sorted(random.sample(range(0, FLAGS.max_seq_length), FLAGS.max_predictions_per_seq)) for i in range(batch_size)])
         masks_list = tf.constant([MASK_ID] * (FLAGS.max_predictions_per_seq * batch_size))
         masked_lm_weights = tf.ones(modeling.get_shape_list(masked_lm_positions))
-        segment_ids = tf.zeros(modeling.get_shape_list(input_ids))
+        segment_ids = tf.zeros(modeling.get_shape_list(input_ids), dtype=tf.int32)
 
         masked_input_ids = replace_elements_by_indices(input_ids, masks_list, masked_lm_positions)
         masked_lm_ids = gather_indexes_rank2(input_ids, masked_lm_positions)
